refactor(client)!: import the generated runtime from OpenAPI.Runtime
Generated clients and server stubs previously carried a pasted copy of the
~2,000-line protocol runtime, while also depending on OpenAPI for the schema
engine. Promote those templates into a real OpenAPI.Runtime module that
generated modules import, so a generated client now contains only its own
document data, models, and typed operations (the single-endpoint corpus case
drops from ~2,400 to ~450 lines).

Document-specific data moves into Runtime.Spec: each generated module packages
its schema resources, security schemes, and default server into a `_SPEC`
constant that its Client values carry, keeping schema-graph caches and server
overrides isolated per module. Public conveniences (Client(), server!,
credential!, ...) are emitted as module-local forwarders rather than methods
on Runtime generics, so independently generated modules can never clobber
each other. Models still extend Runtime._decode/_encode/_form_fields with
methods on their own types.

The HTTP transport core (_request, _stream_request and stream pumping) lives
in OpenAPIHTTPExt as methods on Runtime stubs, preserving OpenAPI's weak HTTP
dependency; generated clients keep `using HTTP`, which loads the extension.
The zoned date-time codecs move to a new OpenAPITimeZonesExt loaded the same
way by `datetime = :zoned` modules.

BREAKING: generated-module `Client` is now a builder function returning
Runtime.Client rather than a module-local struct type, and runtime internals
are no longer defined inside generated modules.
