+++ test/models.jl
@@ -140,12 +140,12 @@
     call(name, args...) = Base.invokelatest(getfield(C, name), args...)
 
     media(name) = only(only(getfield(C, Symbol("_OP_", name)).responses).media)
-    @test call(:_schema_valid, media("truevalue")[3], Dict("anything" => 1))
-    @test !call(:_schema_valid, media("falsevalue")[3], nothing)
-    @test call(:_schema_valid, media("constseven")[3], 7)
-    @test !call(:_schema_valid, media("constseven")[3], 8)
-    @test call(:_schema_valid, media("forbiddenstring")[3], 1)
-    @test !call(:_schema_valid, media("forbiddenstring")[3], "no")
+    @test call(:_schema_valid, C._SPEC, media("truevalue")[3], Dict("anything" => 1))
+    @test !call(:_schema_valid, C._SPEC, media("falsevalue")[3], nothing)
+    @test call(:_schema_valid, C._SPEC, media("constseven")[3], 7)
+    @test !call(:_schema_valid, C._SPEC, media("constseven")[3], 8)
+    @test call(:_schema_valid, C._SPEC, media("forbiddenstring")[3], 1)
+    @test !call(:_schema_valid, C._SPEC, media("forbiddenstring")[3], "no")
 
     @test call(:_decode, C.MixedEnum, "x").value == "x"
     @test call(:_decode, C.MixedEnum, nothing).value === nothing
